backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pinecone==7.3.0
requests>=2.32.3
PyPDF2==3.0.1
python-docx==1.0.1
python-multipart==0.0.9
pydantic==2.11.7
tiktoken==0.11.0
aiofiles==23.2.1
sentence-transformers==3.0.1
openai>=1.35.0
supabase>=2.25.0
python-jose[cryptography]>=3.5.0
openpyxl>=3.1.2